python-app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
azure-storage-blob==12.19.0
azure-communication-email==1.0.0
azure-monitor-opentelemetry==1.2.0
opencensus-ext-azure==1.1.13
reportlab==4.0.7
pillow==10.1.0
requests==2.31.0
email-validator==2.1.0
python-dotenv==1.0.0
cryptography>=41.0.0
aiohttp==3.9.1